registro-minis/requirements.txt
mysql-connector-python==8.3.0
pandas==1.5.3